+++ Editor/Package.cs
@@ -1,8 +1,9 @@
 ﻿
-namespace Hananoki.EditorToolbar {
+namespace HananokiEditor.EditorToolbar {
   public static class Package {
     public const string name = "EditorToolbar";
+    public const string nameNicify = "Editor Toolbar";
     public const string editorPrefName = "Hananoki.EditorToolbar";
-    public const string version = "1.0.9";
+    public const string version = "1.0.10";
   }
 }